+++ enterprise/cmd/frontend/internal/registry/extension_connection_graphql.go
@@ -61,6 +61,7 @@ func toDBExtensionsListOptions(args graphqlbackend.RegistryExtensionConnectionAr
 	if args.PrioritizeExtensionIDs != nil {
 		opt.PrioritizeExtensionIDs = *args.PrioritizeExtensionIDs
 	}
+	opt.ExcludeWIP = !args.IncludeWIP
 	return opt, nil
 }