+++ migrations/1528395559_.down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE registry_extension_releases ALTER COLUMN manifest TYPE text USING manifest#>>'{}';