+++ shared/src/controller.ts
@@ -144,8 +144,10 @@ export class Controller<S extends SettingsSubject, C extends Settings> {
         )
     }
 
-    public withConfiguration(registryExtensions: GQL.IRegistryExtension[]): ConfiguredExtension[] {
-        const configuredExtensions: ConfiguredExtension[] = []
+    public withConfiguration(
+        registryExtensions: GQL.IRegistryExtension[]
+    ): ConfiguredExtension<GQL.IRegistryExtension>[] {
+        const configuredExtensions: ConfiguredExtension<GQL.IRegistryExtension>[] = []
         for (const registryExtension of registryExtensions) {
             configuredExtensions.push({
                 id: registryExtension.extensionID,