+++ web/src/extensions/extension/ExtensionAreaHeader.tsx
@@ -9,6 +9,7 @@ import { isExtensionAdded } from '../ExtensionsClientCommonContext'
 import { ExtensionToggle } from '../ExtensionToggle'
 import { ExtensionAreaRouteContext } from './ExtensionArea'
 import { ExtensionConfigurationState } from './ExtensionConfigurationState'
+import { WorkInProgressBadge } from './WorkInProgressBadge'
 
 interface ExtensionAreaHeaderProps extends ExtensionAreaRouteContext, RouteComponentProps<{}> {
     navItems: ReadonlyArray<ExtensionAreaHeaderNavItem>
@@ -32,6 +33,8 @@ export const ExtensionAreaHeader: React.SFC<ExtensionAreaHeaderProps> = (props: 
     } catch (e) {
         // noop
     }
+
+    const isWorkInProgress = props.extension.registryExtension && props.extension.registryExtension.isWorkInProgress
 
     return (
         <div className="extension-area-header border-bottom simple-area-header">
@@ -60,7 +63,19 @@ export const ExtensionAreaHeader: React.SFC<ExtensionAreaHeaderProps> = (props: 
                                     {manifest &&
                                         manifest.title && <div className="text-muted">{props.extension.id}</div>}
                                     {manifest &&
-                                        manifest.description && <p className="mt-1 mb-0">{manifest.description}</p>}
+                                        (manifest.description || isWorkInProgress) && (
+                                            <p className="mt-1 mb-0">
+                                                {isWorkInProgress && (
+                                                    <WorkInProgressBadge
+                                                        viewerCanAdminister={
+                                                            !!props.extension.registryExtension &&
+                                                            props.extension.registryExtension.viewerCanAdminister
+                                                        }
+                                                    />
+                                                )}
+                                                {manifest.description}
+                                            </p>
+                                        )}
                                 </div>
                             </div>
                         </div>